Fix implicit-const-int-float-conversion warnings https://bugs.webkit.org/show_bug.cgi?id=247559
Reviewed by Michael Catanzaro.

* Source/JavaScriptCore/runtime/DatePrototype.cpp:
(JSC::JSC_DEFINE_HOST_FUNCTION):
"934:93: error: implicit conversion from 'std::numeric_limits<long>::_Ty' (aka 'long') to 'double'
                changes value from 9223372036854775807 to 9223372036854775808"

* Source/WebCore/rendering/style/RenderStyle.cpp:
(WebCore::RenderStyle::setWordSpacing):
"2036:55: error: implicit conversion from 'const int' to 'float' changes value from 33554429 to 33554428"

* Source/WebCore/style/StyleBuilderConverter.h:
(WebCore::Style::BuilderConverter::convertLineHeight):
"1593:68: error: implicit conversion from 'const int' to 'float' changes value from 33554429 to 33554428"

Canonical link: https://commits.webkit.org/256404@main

+++ Source/WebCore/style/StyleBuilderConverter.h
@@ -1590,7 +1590,7 @@ inline std::optional<Length> BuilderConverter::convertLineHeight(BuilderState& b
             length = primitiveValue.computeLength<Length>(conversionData);
         else {
             auto value = primitiveValue.cssCalcValue()->createCalculationValue(conversionData)->evaluate(builderState.style().computedFontSize());
-            length = { clampTo<float>(value, minValueForCssLength, maxValueForCssLength), LengthType::Fixed };
+            length = { clampTo<float>(value, minValueForCssLength, static_cast<float>(maxValueForCssLength)), LengthType::Fixed };
         }
         if (multiplier != 1.f)
             length = Length(length.value() * multiplier, LengthType::Fixed);
